x/exp/ast: move ast into the x/exp package
Addresses IDX-225
