+++ ast/ast_test.go
@@ -6,9 +6,9 @@ import (
 	"time"
 
 	"github.com/cedar-policy/cedar-go/ast"
-	internalast "github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/internal/testutil"
 	"github.com/cedar-policy/cedar-go/types"
+	internalast "github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 func TestASTByTable(t *testing.T) {
