+++ ast/node.go
@@ -1,6 +1,6 @@
 package ast
 
-import "github.com/cedar-policy/cedar-go/internal/ast"
+import "github.com/cedar-policy/cedar-go/x/exp/ast"
 
 // Node is a wrapper type for all the Cedar language operators.  See the [Cedar operators documentation] for details.
 //
