+++ ast/operator.go
@@ -1,8 +1,8 @@
 package ast
 
 import (
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/types"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 //   ____                                 _
