+++ ast/policy.go
@@ -9,9 +9,9 @@ package ast
 import (
 	"bytes"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/internal/json"
 	"github.com/cedar-policy/cedar-go/internal/parser"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 type Policy ast.Policy
