+++ ast/value.go
@@ -4,8 +4,8 @@ import (
 	"net/netip"
 	"time"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/types"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 // Boolean creates a value node containing a Boolean.
