+++ internal/eval/compile.go
@@ -3,8 +3,8 @@ package eval
 import (
 	"fmt"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/types"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 type BoolEvaler struct {
