+++ internal/eval/convert.go
@@ -3,9 +3,9 @@ package eval
 import (
 	"fmt"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/internal/consts"
 	"github.com/cedar-policy/cedar-go/types"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 func toEval(n ast.IsNode) Evaler {
