+++ internal/eval/fold.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"slices"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/types"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 // foldPolicy takes in a given policy and attempts as much constant folding as possible.
