+++ internal/eval/partial.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"slices"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/internal/mapset"
 	"github.com/cedar-policy/cedar-go/types"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 const variableEntityType = "__cedar::variable"
