+++ internal/json/json_test.go
@@ -5,9 +5,9 @@ import (
 	"net/netip"
 	"testing"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/internal/testutil"
 	"github.com/cedar-policy/cedar-go/types"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 func TestUnmarshalJSON(t *testing.T) {
