+++ internal/parser/cedar_marshal.go
@@ -4,9 +4,9 @@ import (
 	"bytes"
 	"fmt"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/internal/consts"
 	"github.com/cedar-policy/cedar-go/internal/extensions"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 func (p *Policy) MarshalCedar(buf *bytes.Buffer) {
