+++ internal/parser/cedar_unmarshal.go
@@ -5,11 +5,11 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/internal/consts"
 	"github.com/cedar-policy/cedar-go/internal/extensions"
 	"github.com/cedar-policy/cedar-go/internal/mapset"
 	"github.com/cedar-policy/cedar-go/types"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 func (p *PolicySlice) UnmarshalCedar(b []byte) error {
