+++ internal/parser/internal_test.go
@@ -3,8 +3,8 @@ package parser
 import (
 	"testing"
 
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/internal/testutil"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 func TestScopeToNode(t *testing.T) {
