+++ internal/parser/policy.go
@@ -1,6 +1,6 @@
 package parser
 
-import "github.com/cedar-policy/cedar-go/internal/ast"
+import "github.com/cedar-policy/cedar-go/x/exp/ast"
 
 type PolicySlice []*Policy
 type Policy ast.Policy
