+++ x/exp/batch/batch.go
@@ -14,11 +14,11 @@ import (
 	"slices"
 
 	"github.com/cedar-policy/cedar-go"
-	"github.com/cedar-policy/cedar-go/internal/ast"
 	"github.com/cedar-policy/cedar-go/internal/consts"
 	"github.com/cedar-policy/cedar-go/internal/eval"
 	"github.com/cedar-policy/cedar-go/internal/mapset"
 	"github.com/cedar-policy/cedar-go/types"
+	"github.com/cedar-policy/cedar-go/x/exp/ast"
 )
 
 // Ignore returns a value that should be ignored during batch evaluation.
